Remove obsolete projects and add new browser tests
Significant cleanup and removal of various projects and files:
- Removed `Chameleon.lib.Auth.csproj`, `Oidc.browser.cs`, `Oidc.client.cs`, `Session.cs`, `Chameleon - Backup.lib.CommunityToolkit.csproj`, `Chameleon.lib.CommunityToolkit.csproj`, `ObservableObjectBase.cs`, `ViewModelObjectBase.cs`, `Vim.cs`, `AuthTests.cs`, `Base.cs`, `Chameleon.lib.Tests.csproj`, `IOCTest.cs`, `Playutilstests.cs`, and related test files.
- Removed all Playwright integration tests and configurations, including `PlaywrightIntegrationTests.cs`, `PlaywrightTestRunnerTests.cs`, `PlaywrightTestsBase.cs`, `BrowserLauncherTests.cs`, `ExtensionLoaderServiceTests.cs`, `WebBroswserTestsBase.cs`, and `appsettings.json`.

Updated solution and added new test functionalities:
- Updated `Chameleon.lib.sln` to remove obsolete projects and add `Chameleon.lib.WebBrowser`.
- Added new test classes: `IOCTest`, `AuthTests`, `ApiTestsBase`, `Playutilstests`, `PlaywrightIntegrationTests`, `PlaywrightTestRunnerTests`.
- Introduced new classes for browser-related tests: `PlaywrightTestsBase`, `BrowserLauncherTests`, `ExtensionLoaderServiceTests`, `WebBroswserTestsBase`.
- Enhanced test coverage and modularity with dependency injection and configuration setup using the IoC container.

+++ Tests/WebBrowser/ExtensionLoaderServiceTests.cs
@@ -1,17 +1,13 @@
-﻿using System;
-using System.Diagnostics;
-using Chameleon.lib.WebBrowser.Services;
+﻿using System.Diagnostics;
 
-using Microsoft.Extensions.Configuration;
-using Microsoft.Extensions.DependencyInjection;
+using Chameleon.lib.WebBrowser.Services;
 
 using static Chameleon.lib.Common.Constants.Enums;
 
-namespace Chameleon.lib.Tests.WebBrowser;
+namespace Tests.WebBrowser;
 public class ExtensionLoaderServiceTests : WebBroswserTestsBase {
 	[Fact]
-	public async Task LoadExtension_ValidExtension_Succeeds()
-	{
+	public async Task LoadExtension_ValidExtension_Succeeds() {
 		_ = await _tcs.Task;
 
 		// Arrange
@@ -42,9 +38,7 @@ public class ExtensionLoaderServiceTests : WebBroswserTestsBase {
 
 		} finally {
 			// Clean up
-			if (Directory.Exists(destinationPath)) {
-				Directory.Delete(destinationPath, true);
-			}
+			if (Directory.Exists(destinationPath)) Directory.Delete(destinationPath, true);
 		}
 	}
 }
